refactor: hooks 폴더 분류에 따른 코드 변경 수정
