+++ src/hooks/common/useMediaQuery.ts
@@ -1,6 +1,6 @@
 import { useEffect, useMemo, useState } from 'react';
 
-import { SCREEN_BREAKPOINT } from '../constants/styles';
+import { SCREEN_BREAKPOINT } from '../../constants/styles';
 
 type ScreenKey = `is${Capitalize<keyof typeof SCREEN_BREAKPOINT>}`;
 type Screen = Map<ScreenKey, MediaQueryList>;
