+++ src/pages/Messages/index.tsx
@@ -5,7 +5,7 @@ import { MAIN_MESSAGES_KEY, getMessageLists } from '../../api/messages';
 
 import MessageList from '../../components/messages/MessagesList';
 import { fetchQuery } from '../../utils/fetchQuery';
-import useInfinityObserver from '../../hooks/useInfinityObserver';
+import useInfinityObserver from '../../hooks/common/useInfinityObserver';
 import { userStore } from '../../stores/userStore';
 
 const MessagesPage = () => {
