--- src/pages/Test/index.tsx
@@ -1,10 +1,0 @@
-const TestPage = () => {
-  return (
-    <>
-      <div>Modal 테스트 페이지!</div>
-      <button type="button">삭제</button>
-    </>
-  );
-};
-
-export default TestPage;
